requirements.txt
python-dotenv
cryptography
web3
solders
base58